Move eithercart functions to yoke module
